refactor(recording settings): add DesktopOutputFolder component for managing output folder selection
