+++ apps/whispering/src/routes/(config)/settings/recording/+page.svelte
@@ -1,15 +1,13 @@
 <script lang="ts">
+  import DesktopOutputFolder from './DesktopOutputFolder.svelte';
 	import { LabeledSelect } from '$lib/components/labeled/index.js';
 	import { Separator } from '@repo/ui/separator';
-	import { Button } from '@repo/ui/button';
-	import { FolderOpen, ExternalLink } from 'lucide-svelte';
 	import {
 		BITRATE_OPTIONS,
 		RECORDING_MODE_OPTIONS,
 	} from '$lib/constants/audio';
 	import { settings } from '$lib/stores/settings.svelte';
 	import SelectRecordingDevice from './SelectRecordingDevice.svelte';
-	import { asDeviceIdentifier } from '$lib/services/types';
 
 	const SAMPLE_RATE_OPTIONS = [
 		{ value: '16000', label: 'Voice Quality (16kHz): Optimized for speech' },
@@ -17,33 +15,6 @@
 		{ value: '48000', label: 'High Quality (48kHz): Professional audio' },
 	] as const;
 
-	async function selectOutputFolder() {
-		if (!window.__TAURI_INTERNALS__) return;
-		
-		const { open } = await import('@tauri-apps/plugin-dialog');
-		const selected = await open({
-			directory: true,
-			multiple: false,
-			title: 'Select Recording Output Folder',
-		});
-		
-		if (selected) settings.updateKey('recording.desktop.outputFolder', selected);
-	}
-
-	async function openOutputFolder() {
-		if (!window.__TAURI_INTERNALS__) return;
-		const { openPath } = await import('@tauri-apps/plugin-opener');
-		
-		const folderPath = settings.value['recording.desktop.outputFolder'];
-		if (!folderPath) {
-			// If no custom folder is set, get the app data directory
-			const { appDataDir } = await import('@tauri-apps/api/path');
-			const defaultPath = await appDataDir();
-			await openPath(defaultPath);
-		} else {
-			await openPath(folderPath);
-		}
-	}
 </script>
 
 <svelte:head>
@@ -115,35 +86,7 @@
 				<label for="output-folder" class="text-sm font-medium">
 					Recording Output Folder
 				</label>
-				<div class="flex gap-2">
-					<div class="flex-1 px-3 py-2 text-sm border rounded-md bg-muted">
-						{settings.value['recording.desktop.outputFolder'] ?? 'Default (App Data Folder)'}
-					</div>
-					<Button
-						variant="outline"
-						onclick={selectOutputFolder}
-					>
-						<FolderOpen class="w-4 h-4 mr-2" />
-						Browse
-					</Button>
-					<Button
-						variant="outline"
-						onclick={openOutputFolder}
-					>
-						<ExternalLink class="w-4 h-4 mr-2" />
-						Open
-					</Button>
-					{#if settings.value['recording.desktop.outputFolder']}
-						<Button
-							variant="outline"
-							onclick={() => {
-								settings.updateKey('recording.desktop.outputFolder', null);
-							}}
-						>
-							Reset
-						</Button>
-					{/if}
-				</div>
+				<DesktopOutputFolder></DesktopOutputFolder>
 				<p class="text-xs text-muted-foreground">
 					Choose where to save your recordings. Default location is secure and managed by the app.
 				</p>
